requirements.txt
aniso8601==8.0.0
astroid==2.4.2
click==7.1.2
Flask==1.1.2
Flask-RESTful==0.3.8
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
psycopg2-binary==2.8.5
pylint==2.5.3
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
Werkzeug==1.0.1
wrapt==1.12.1
